+++ framework/db/QueryBuilder.php
@@ -176,7 +176,7 @@ class QueryBuilder extends \yii\base\Object
         $placeholders = [];
         $values = ' DEFAULT VALUES';
         if ($columns instanceof \yii\db\Query) {
-            list($names, $values) = $this->prepareInsertSelectSubQuery($columns, $schema);
+            list($names, $values, $params) = $this->prepareInsertSelectSubQuery($columns, $schema);
         } else {
             foreach ($columns as $name => $value) {
                 $names[] = $schema->quoteColumnName($name);
@@ -206,16 +206,19 @@ class QueryBuilder extends \yii\base\Object
      *
      * @param \yii\db\Query $columns Object, which represents select query
      * @param \yii\db\Schema $schema Schema object to qoute column name
+     * @param array $params the parameters to be bound to the generated SQL statement. These parameters will
+     * be included in the result with the additional parameters generated during the query building process.
      * @return array
+     * @throws InvalidParamException if query's select does not contain only named parameters.
      * @since 2.0.11
      */
-    protected function prepareInsertSelectSubQuery($columns, $schema)
+    protected function prepareInsertSelectSubQuery($columns, $schema, $params = [])
     {
         if (!is_array($columns->select) || empty($columns->select) || in_array('*', $columns->select)) {
             throw new InvalidParamException('Expected select query object with enumerated (named) parameters');
         }
 
-        list ($values, ) = $this->build($columns);
+        list ($values, $params) = $this->build($columns, $params);
         $names = [];
         $values = ' ' . $values;
         foreach ($columns->select as $title => $field) {
@@ -228,7 +231,7 @@ class QueryBuilder extends \yii\base\Object
             }
         }
 
-        return [$names, $values];
+        return [$names, $values, $params];
     }
 
     /**